refactor: move tree util functions to utils folder
